Add arch filtering, restructure pkg
