+++ index/server/pkg/server/registry.go
@@ -1,0 +1,109 @@
+package server
+
+import (
+	"context"
+	"errors"
+	"fmt"
+	"io/ioutil"
+	"log"
+	"path"
+	"path/filepath"
+
+	indexSchema "github.com/devfile/registry-support/index/generator/schema"
+
+	"github.com/containerd/containerd/remotes/docker"
+	"github.com/deislabs/oras/pkg/content"
+	"github.com/deislabs/oras/pkg/oras"
+	ocispec "github.com/opencontainers/image-spec/specs-go/v1"
+)
+
+// pushStackToRegistry pushes the given devfile stack to the OCI registry
+func pushStackToRegistry(devfileIndex indexSchema.Schema) error {
+	// Load the devfile into memory and set up the pushing resource (file name, file content, media type, ref)
+	memoryStore := content.NewMemoryStore()
+	pushContents := []ocispec.Descriptor{}
+	for _, resource := range devfileIndex.Resources {
+		if resource == "meta.yaml" {
+			// Some registries may still have the meta.yaml in it, but we don't need it, so skip pushing it up
+			continue
+		}
+
+		// Get the media type that corresponds to the resource
+		// Some resources have media types that depends on the entire filename (e.g. devfile.yaml, archive.tar),
+		// others just depend on the file extension (e.g. vsx files)
+		var mediaType string
+		var found bool
+		switch resource {
+		case devfileName, devfileNameHidden, svgLogoName, pngLogoName, archiveName:
+			// Get the media type associated with the file
+			if mediaType, found = mediaTypeMapping[resource]; !found {
+				return errors.New("media type not found for file " + resource)
+			}
+		default:
+			// Probably vsx file, but get the extension of the file just in case
+			fileExtension := filepath.Ext(resource)
+			if mediaType, found = mediaTypeMapping[fileExtension]; !found {
+				return errors.New("media type not found for file extension" + fileExtension)
+			}
+		}
+
+		// Load the resource into memory and add to the push contents
+		resourceContent, err := ioutil.ReadFile(filepath.Join(stacksPath, devfileIndex.Name, resource))
+		if err != nil {
+			return err
+		}
+
+		desc := memoryStore.Add(resource, mediaType, resourceContent)
+		pushContents = append(pushContents, desc)
+
+	}
+
+	ref := path.Join(registryService, "/", devfileIndex.Links["self"])
+
+	ctx := context.Background()
+	resolver := docker.NewResolver(docker.ResolverOptions{PlainHTTP: true})
+	log.Printf("Pushing %s to %s...\n", devfileIndex.Name, ref)
+	desc, err := oras.Push(ctx, resolver, ref, memoryStore, pushContents, oras.WithConfigMediaType(devfileConfigMediaType))
+	if err != nil {
+		return fmt.Errorf("failed to push %s to %s: %v", devfileIndex.Name, ref, err)
+	}
+	log.Printf("Pushed to %s with digest %s\n", ref, desc.Digest)
+	return nil
+}
+
+// pullStackFromRegistry pulls the given devfile stack from the OCI registry
+func pullStackFromRegistry(devfileIndex indexSchema.Schema) ([]byte, error) {
+	// Pull the devfile from registry and save to disk
+	ref := path.Join(registryService, "/", devfileIndex.Links["self"])
+
+	ctx := context.Background()
+	resolver := docker.NewResolver(docker.ResolverOptions{PlainHTTP: true})
+
+	// Initialize memory store
+	memoryStore := content.NewMemoryStore()
+	allowedMediaTypes := []string{devfileMediaType}
+
+	var devfile string
+	for _, resource := range devfileIndex.Resources {
+		if resource == devfileName {
+			devfile = devfileName
+			break
+		}
+		if resource == devfileNameHidden {
+			devfile = devfileNameHidden
+			break
+		}
+	}
+	log.Printf("Pulling %s from %s...\n", devfile, ref)
+	desc, _, err := oras.Pull(ctx, resolver, ref, memoryStore, oras.WithAllowedMediaTypes(allowedMediaTypes))
+	if err != nil {
+		return nil, fmt.Errorf("failed to pull %s from %s: %v", devfile, ref, err)
+	}
+	_, bytes, ok := memoryStore.GetByName(devfile)
+	if !ok {
+		return nil, fmt.Errorf("failed to load %s to memory", devfile)
+	}
+
+	log.Printf("Pulled from %s with digest %s\n", ref, desc.Digest)
+	return bytes, nil
+}
