+++ tests/registry/stacks/java-maven/devfile.yaml
@@ -1,4 +1,4 @@
-schemaVersion: 2.0.0
+schemaVersion: 2.1.0
 metadata:
   name: java-maven
   version: 1.1.0
@@ -7,6 +7,10 @@ metadata:
   tags: ["Java", "Maven"]
   projectType: "maven"
   language: "java"
+  architectures:
+    - "amd64"
+    - "arm64"
+    - "s390x"
 starterProjects:
   - name: springbootproject
     git:
